feat: enable PPO+EKF training with checkpoint resume and student eval
- Add EKF wrapping to RL training loop (training/loops/rl.py)
- Fix action space dtype float64→float32 for SB3 compatibility
- Add playground_ekf_ppo experiment config
- Add DAgger student evaluation script
- Add tests for EKF RL loop and student evaluation

+++ sim/envs/rate_ctrl_env.py
@@ -177,7 +177,7 @@ class RateCtrlEnv:
             low=-np.inf, high=np.inf, shape=(OBS_DIM,), dtype=np.float64,
         )
         self.action_space = spaces.Box(
-            low=0.0, high=1.0, shape=(ACT_DIM,), dtype=np.float64,
+            low=0.0, high=1.0, shape=(ACT_DIM,), dtype=np.float32,
         )
 
         # Allocate persistent arrays
